Fixes for running fi_pingpong over in-kernel-ses on mips - 1. Temporary fix for UET demo app posting recv after receiving SOM from intiator sometimes. 2. Changing rx_pkt callback prototype of uet_nic_ops (pkt_len - from size_t to int). 3. Setting netns_ok of uet_protocol (struct net_protocol) for compiling with linux_3.10. 4. Adding some debug prints in uet_parse_pkt (TODO: pr_info needs to be replaced with pr_debug).

+++ in-kernel-ses/driver/uet_core/uet_nic.c
@@ -24,7 +24,7 @@ static int (*uet_nic_tx_pkt_fn)(struct uet_nic *nic,
 static int (*uet_nic_rx_pkt_fn)(struct uet_nic *nic,
 			      void *pkt,
 			      size_t pkt_buf_size,
-			      size_t *rx_pkt_size);
+			      int *rx_pkt_size);
 static int (*uet_nic_rx_poll_fn)(struct uet_nic *nic);
 static int (*uet_nic_mr_reg_fn)(struct uet_nic *nic,
 			      struct uet_mr_buf_desc *desc,
